Fix an obvious error that would cause fpu_ftoxi() to drop the upper half. The compiler even emitted a warning for this...

+++ sys/arch/sparc/fpu/fpu_implode.c
@@ -1,4 +1,4 @@
-/*	$NetBSD: fpu_implode.c,v 1.5 2000/06/18 06:54:17 mrg Exp $ */
+/*	$NetBSD: fpu_implode.c,v 1.6 2000/07/24 04:13:45 mycroft Exp $ */
 
 /*
  * Copyright (c) 1992, 1993
@@ -272,7 +272,7 @@ fpu_ftoxi(fe, fp, res)
 		/* NB: the following includes exp < 0 cases */
 		if (fpu_shr(fp, FP_NMANT - 1 - exp) != 0)
 			fe->fe_cx |= FSR_NX;
-		i = (fp->fp_mant[2]<<32)|fp->fp_mant[3];
+		i = ((u_int64_t)fp->fp_mant[2]<<32)|fp->fp_mant[3];
 		if (i >= ((u_int64_t)0x8000000000000000LL + sign))
 			break;
 		return (sign ? -i : i);
